feat!: support many page sections & images
This completely reconfigures the files in `i18n` to support multiple
page sections, so that projects that want more than just the hero
section can add them.

It also moves `src/images` to the root, and handles them correctly! This
took a lot of work. While getting there, it was also useful to add a
GraphQL extension to VS Code and generate the schema it needs using a
Gatsby plugin. I also found a Gatsby plugin that auto-generates types
for all the GraphQL queries we have in our components.