feat: switch to new testnet contract
this copies in the TenK TS file from here, with one modification:
https://raw.githubusercontent.com/TENK-DAO/tenk/feat!/update_api/contracts/tenk/ts/index.ts

the modification: the Status enum is not yet encoded correctly by wit /
NEAR TS Wit Bindgen (NETS WITB). In the file linked above, it sets the
runtime value of each option to a number. In the code here, I've changed
them to strings which match the option name.