feat: use real CSS background images
Gatsby makes this, like so many other things, way too hard.

This adds a community-supported plugin, [gatsby-background-image]. To
make it work, it also moves the image lookup logic to a custom hook,
`useImageData`, so that the same lookup can be used by the regular
foreground images as well as the new background-image component.

On the plus side, this greatly simplifies the `Section` CSS and makes
the styles more robust, meaning they should work in an expectable way in
a much larger range of circumstances.

  [gatsby-background-image]: https://github.com/timhagn/gatsby-background-image/tree/main/packages/gatsby-background-image