fix: s/mintLimit/remainingAllowance/; fix logic
remainingAllowance is a better name, since it's for the premint and
that's generally how the team refers to it.

This value can also be null/undefined for a given user, which is
meaningful. This is now preserved in the `useTenk` logic.

When setting a `max` on the `setNumber` input, we now prefer
`remainingAllowance` if it is not undefined, otherwise use
`mintRateLimit`.